% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralization_function.R
\name{append.CTR.measures}
\alias{append.CTR.measures}
\title{Append computed measures to ouptut data table.}
\usage{
append.CTR.measures(
  measures,
  i,
  path = "output/undirected/centralization_measures.csv"
)
}
\arguments{
\item{measures}{Network measure data table.}

\item{i}{Network index}

\item{path}{Location folder of data table.}
}
\value{
data table
}
\description{
Append computed measures to ouptut data table.
}
