% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfnetzschleuder_functions.R,
%   R/netzschleuder_functions.R
\name{append.NS.measures}
\alias{append.NS.measures}
\title{Append computed measures to ouptut data table.}
\usage{
append.NS.measures(measures, i, path = "output/netzschleuder_measures.csv")

append.NS.measures(measures, i, path = "output/netzschleuder_measures.csv")
}
\arguments{
\item{measures}{Network measure data table.}

\item{i}{Network index}

\item{path}{Location folder of data table.}
}
\value{
data table

data table
}
\description{
Append computed measures to ouptut data table.

Append computed measures to ouptut data table.
}
