% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfOLP_functions.R, R/OLP_functions.R
\name{append.OLP.measures}
\alias{append.OLP.measures}
\title{Append computed measures to ouptut data table.}
\usage{
append.OLP.measures(measures, i, path = "output/OLP_measures.csv")

append.OLP.measures(measures, i, path = "output/OLP_measures.csv")
}
\arguments{
\item{measures}{Network measure data table.}

\item{i}{Network index}

\item{path}{Location folder of data table.}
}
\value{
data table

data table
}
\description{
Append computed measures to ouptut data table.

Append computed measures to ouptut data table.
}
