% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralization_function.R
\name{compute.CTR.measures}
\alias{compute.CTR.measures}
\title{Compute several network measures of a complex network.}
\usage{
\method{compute}{CTR.measures}(
  igraph.network,
  i,
  data = fread("output/undirected/master_measures_2.csv")[, Name]
)
}
\arguments{
\item{igraph.network}{Network as an igraph object.}

\item{i}{Network index}
}
\value{
ID, Network Name, Reciprocity, Degree Distribution, Transitivity, Degree Assortativity, Betweenness, Closeness,
Average Path Lenght, Hierarchy, Density
}
\description{
Compute several network measures of a complex network.
}
