% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfICON_functions.R, R/ICON_functions.R
\name{compute.ICON.measures}
\alias{compute.ICON.measures}
\title{Compute several network measures of a complex network.}
\usage{
compute.ICON.measures(
  igraph.network,
  i,
  ICON_data = fread("input/import_datasets/ICON_data.csv")
)

compute.ICON.measures(
  igraph.network,
  i,
  ICON_data = fread("input/import_datasets/ICON_data.csv")
)
}
\arguments{
\item{igraph.network}{Network as an igraph object.}

\item{i}{Network index}
}
\value{
ID, Network Name, Reciprocity, Degree Distribution, Transitivity, Degree Assortativity, Betweenness, Closeness,
Average Path Lenght, Hierarchy, Density

ID, Network Name, Reciprocity, Degree Distribution, Transitivity, Degree Assortativity, Betweenness, Closeness,
Average Path Lenght, Hierarchy, Density
}
\description{
Compute several network measures of a complex network.

Compute several network measures of a complex network.
}
