% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLP_functions.R
\name{compute.OLP.measures}
\alias{compute.OLP.measures}
\title{Compute several network measures of a complex network.}
\usage{
compute.OLP.measures(
  igraph.network,
  i,
  OLP_data = fread("input/import_datasets/OLP_essentials.csv"),
  directed = OLP_data[i, directed]
)
}
\arguments{
\item{igraph.network}{Network as an igraph object.}

\item{i}{Network index}
}
\value{
ID, Network Name, Network Domain, Reciprocity, Degree Distribution, Transitivity, Degree Assortativity, Betweenness, Closeness,
Average Path Lenght, Hierarchy, Density
}
\description{
Compute several network measures of a complex network.
}
