% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralization_function.R
\name{create.igraph.object.CTR}
\alias{create.igraph.object.CTR}
\title{Convert .csv object to igraph object}
\usage{
create.igraph.object.CTR(network, i)
}
\arguments{
\item{network}{Network name as a .csv object (corresponds to Name in master_measures)}
}
\value{
igraph object
}
\description{
Convert .csv object to igraph object
}
