% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfICON_functions.R, R/ICON_functions.R
\name{create.igraph.object.ICON}
\alias{create.igraph.object.ICON}
\title{Convert .rda object to igraph object}
\usage{
create.igraph.object.ICON(
  network,
  i,
  ICON_data = fread("input/import_datasets/ICON_data.csv")
)

create.igraph.object.ICON(
  network,
  i,
  ICON_data = fread("input/import_datasets/ICON_data.csv")
)
}
\arguments{
\item{network}{Network name as .csv object (corresponds to var_name in ICON_data)}
}
\value{
igraph object

igraph object
}
\description{
Convert .rda object to igraph object

Convert .rda object to igraph object
}
