% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfnetzschleuder_functions.R,
%   R/netzschleuder_functions.R
\name{create.igraph.object.NS}
\alias{create.igraph.object.NS}
\title{Convert .csv object to igraph object}
\usage{
create.igraph.object.NS(
  network,
  i,
  netzschleuder_data = fread("input/import_datasets/netzschleuder_essentials.csv")
)

create.igraph.object.NS(
  network,
  i,
  netzschleuder_data = fread("input/import_datasets/netzschleuder_essentials.csv")
)
}
\arguments{
\item{network}{Network name as a .csv object (corresponds to network_name)}
}
\value{
igraph object

igraph object
}
\description{
Convert .csv object to igraph object

Convert .csv object to igraph object
}
