% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLP_functions.R
\name{create.igraph.object.OLP}
\alias{create.igraph.object.OLP}
\title{Convert .csv object to igraph object}
\usage{
create.igraph.object.OLP(network)
}
\arguments{
\item{network}{Network name as a .csv object (corresponds to network_name)}
}
\value{
igraph object
}
\description{
Convert .csv object to igraph object
}
