% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfICON_functions.R, R/ICON_functions.R
\name{ICON.network.measures}
\alias{ICON.network.measures}
\title{Compute and append network measures to data table.}
\usage{
ICON.network.measures(network, i, path = "output/ICON_measures.csv")

ICON.network.measures(network, i, path = "output/ICON_measures.csv")
}
\arguments{
\item{network}{Network name as an .rda object (corresponds to var_name in ICON_data)}

\item{i}{Network index}

\item{path}{Location folder of data table.}
}
\value{
data table

data table
}
\description{
Compute and append network measures to data table.

Compute and append network measures to data table.
}
