% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfnetzschleuder_functions.R,
%   R/netzschleuder_functions.R
\name{NS.network.measures}
\alias{NS.network.measures}
\title{Compute and append network measures to data table.}
\usage{
NS.network.measures(network, i, path = "output/netzschleuder_measures.csv")

NS.network.measures(network, i, path = "output/netzschleuder_measures.csv")
}
\arguments{
\item{network}{Network name as an .rda object (corresponds to var_name in ICON_data)}

\item{i}{Network index}

\item{path}{Location folder of data table.}
}
\value{
data table

data table
}
\description{
Compute and append network measures to data table.

Compute and append network measures to data table.
}
