% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLP_functions.R
\name{OLP.network.measures}
\alias{OLP.network.measures}
\title{Compute and append network measures to data table.}
\usage{
OLP.network.measures(network, i, path = "output/OLP_measures.csv")
}
\arguments{
\item{network}{Network name as an .rda object (corresponds to var_name in ICON_data)}

\item{i}{Network index}

\item{path}{Location folder of data table.}
}
\value{
data table
}
\description{
Compute and append network measures to data table.
}
