% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion_function.R
\name{simulate.diffusion}
\alias{simulate.diffusion}
\title{Simulate diffusion on network}
\usage{
\method{simulate}{diffusion}(
  j,
  p.infection,
  pct.starting.infected,
  runs = 1e+06,
  master_data = fread("output/undirected/master_measures_2.csv")[, Name]
)
}
\arguments{
\item{j}{network index (corresponds to Name column in master_data)}

\item{p.infection}{probability of infection}

\item{pct.starting.infected}{percentage of nodes starting infected}
}
\value{
number of iterations for 70% of nodes infected
}
\description{
Simulate diffusion on network
}
