% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_diffusion_function.R
\name{simulate.removed.diffusion}
\alias{simulate.removed.diffusion}
\title{Simulate diffusion on network}
\usage{
\method{simulate}{removed.diffusion}(
  j,
  p.infection,
  pct.starting.infected,
  n,
  threshold,
  runs = 10000,
  master_data = fread("output/undirected/master_measures_2.csv")[, c("Name",
    "NetworkDomain", "Edges", "Nodes")]
)
}
\arguments{
\item{j}{network index (corresponds to 'Name' column in master_data)}

\item{p.infection}{probability of infection}

\item{pct.starting.infected}{percentage of nodes starting infected}

\item{n}{number of repetitions of diffusion}
}
\value{
number of infection iterations for 70% of nodes infected
}
\description{
Simulate diffusion on network
}
