% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzschleuder_functions.R
\name{compute.NS.measures}
\alias{compute.NS.measures}
\title{Compute several network measures of a complex network.}
\usage{
compute.NS.measures(
  igraph.network,
  i,
  netzschleuder_data = fread("input/import_datasets/netzschleuder_essentials.csv"),
  directed = netzschleuder_data[i, directed]
)
}
\arguments{
\item{igraph.network}{Network as an igraph object.}

\item{i}{Network index}
}
\value{
ID, Network Name, Reciprocity, Degree Distribution, Transitivity, Degree Assortativity, Betweenness, Closeness,
Average Path Lenght, Hierarchy, Density
}
\description{
Compute several network measures of a complex network.
}
