% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLP_functions.R
\name{convert.OLP}
\alias{convert.OLP}
\title{Convert OLP network data into edgelist}
\usage{
convert.OLP(OLP.network)
}
\arguments{
\item{OLP.network}{Optimal Link Prediction network as list}
}
\value{
data.frame
}
\description{
Convert OLP network data into edgelist
}
