% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICON_functions.R
\name{create.igraph.object.ICON}
\alias{create.igraph.object.ICON}
\title{Convert .rda object to igraph object}
\usage{
create.igraph.object.ICON(network)
}
\arguments{
\item{network}{Network name as an .rda object (corresponds to var_name in ICON_data)}
}
\value{
igraph object
}
\description{
Convert .rda object to igraph object
}
