% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfOLP_functions.R, R/OLP_functions.R
\name{create.igraph.object.OLP}
\alias{create.igraph.object.OLP}
\title{Convert .csv object to igraph object}
\usage{
create.igraph.object.OLP(
  network,
  i,
  OLP_data = fread("input/import_datasets/OLP_essentials.csv")
)

create.igraph.object.OLP(
  network,
  i,
  OLP_data = fread("input/import_datasets/OLP_essentials.csv")
)
}
\arguments{
\item{network}{Network name as a .csv object (corresponds to network_name)}
}
\value{
igraph object

igraph object
}
\description{
Convert .csv object to igraph object

Convert .csv object to igraph object
}
